pyproject.toml
[project]
name = "harp"
version = "1.0.0"
description = "Replication package: Global Persistence, Local Residual Structure"
requires-python = ">=3.11,<3.13"
license = {text = "MIT"}
authors = [{name = "Oleg Roshka"}]

dependencies = [
  "numpy>=1.24,<3.0",
  "pandas>=2.0",
  "scipy>=1.11",
  "scikit-learn>=1.3",
  "pyarrow>=14.0",
  "statsmodels>=0.14",
  "matplotlib>=3.7",
]

[project.optional-dependencies]
gpu = ["torch>=2.2"]
data-pipeline = [
  "yfinance>=0.2",
  "fredapi>=0.5",
  "pydantic>=2.7",
  "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/harp"]
